# perf(serde_v8): zero-copy StringOrBuffer