--- a/serde_v8/Cargo.toml
+++ b/serde_v8/Cargo.toml
@@ -15,7 +15,6 @@ path = "lib.rs"
 [dependencies]
 derive_more = "0.99.17"
 serde = { version = "1.0.130", features = ["derive"] }
-serde_bytes = "0.11"
 v8 = "0.42.0"
 
 [dev-dependencies]