--- a/serde_v8/benches/de.rs
+++ b/serde_v8/benches/de.rs
@@ -155,6 +155,54 @@ fn de_bstr_v8_1024_b(b: &mut Bencher) {
   );
 }
 
+fn de_sob_str_6b(b: &mut Bencher) {
+  dedo("'byebye'", |scope, v| {
+    b.iter(move || {
+      let _: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
+    })
+  })
+}
+
+fn de_sob_str_1kb(b: &mut Bencher) {
+  dedo("'deno'.repeat(256)", |scope, v| {
+    b.iter(move || {
+      let _: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
+    })
+  })
+}
+
+fn de_sob_buf_1b(b: &mut Bencher) {
+  dedo("new Uint8Array([97])", |scope, v| {
+    b.iter(move || {
+      let _: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
+    })
+  })
+}
+
+fn de_sob_buf_1kb(b: &mut Bencher) {
+  dedo("(new Uint8Array(1*1024)).fill(42)", |scope, v| {
+    b.iter(move || {
+      let _: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
+    })
+  })
+}
+
+fn de_sob_buf_16kb(b: &mut Bencher) {
+  dedo("(new Uint8Array(16*1024)).fill(42)", |scope, v| {
+    b.iter(move || {
+      let _: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
+    })
+  })
+}
+
+fn de_sob_buf_512kb(b: &mut Bencher) {
+  dedo("(new Uint8Array(512*1024)).fill(42)", |scope, v| {
+    b.iter(move || {
+      let _: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
+    })
+  })
+}
+
 benchmark_group!(
   benches,
   de_struct_v8,
@@ -174,6 +222,12 @@ benchmark_group!(
   de_tuple_v8_opt,
   de_bstr_v8_12_b,
   de_bstr_v8_1024_b,
+  de_sob_str_6b,
+  de_sob_str_1kb,
+  de_sob_buf_1b,
+  de_sob_buf_1kb,
+  de_sob_buf_16kb,
+  de_sob_buf_512kb,
 );
 
 benchmark_main!(benches);