--- a/serde_v8/magic/string_or_buffer.rs
+++ b/serde_v8/magic/string_or_buffer.rs
@@ -1,45 +1,50 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+use super::buffer::MagicBuffer;
+use super::transl8::{FromV8, ToV8};
+use crate::magic::transl8::impl_magic;
+use crate::Error;
 use std::ops::Deref;
 
 #[derive(Debug)]
-pub struct StringOrBuffer(Vec<u8>);
-
-impl Deref for StringOrBuffer {
-  type Target = Vec<u8>;
-  fn deref(&self) -> &Vec<u8> {
-    &self.0
-  }
-}
-
-impl StringOrBuffer {
-  pub fn into_bytes(self) -> Vec<u8> {
-    self.0
-  }
-}
-
-impl<'de> serde::Deserialize<'de> for StringOrBuffer {
-  fn deserialize<D>(deserializer: D) -> Result<StringOrBuffer, D::Error>
-  where
-    D: serde::Deserializer<'de>,
-  {
-    StringOrBufferInner::deserialize(deserializer)
-      .map(|x| StringOrBuffer(x.into_bytes()))
-  }
-}
-
-// TODO(@AaronO): explore if we can make this work with ZeroCopyBuf
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum StringOrBufferInner {
-  #[serde(with = "serde_bytes")]
-  Buffer(Vec<u8>),
+pub enum StringOrBuffer {
+  Buffer(MagicBuffer),
   String(String),
 }
 
-impl StringOrBufferInner {
-  fn into_bytes(self) -> Vec<u8> {
+impl_magic!(StringOrBuffer);
+
+impl Deref for StringOrBuffer {
+  type Target = [u8];
+  fn deref(&self) -> &Self::Target {
     match self {
-      Self::String(s) => s.into_bytes(),
-      Self::Buffer(b) => b,
+      Self::Buffer(b) => b.as_ref(),
+      Self::String(s) => s.as_bytes(),
     }
+  }
+}
+
+impl ToV8 for StringOrBuffer {
+  fn to_v8<'a>(
+    &self,
+    scope: &mut v8::HandleScope<'a>,
+  ) -> Result<v8::Local<'a, v8::Value>, crate::Error> {
+    match self {
+      Self::Buffer(buf) => crate::to_v8(scope, buf),
+      Self::String(s) => crate::to_v8(scope, s),
+    }
+  }
+}
+
+impl FromV8 for StringOrBuffer {
+  fn from_v8(
+    scope: &mut v8::HandleScope,
+    value: v8::Local<v8::Value>,
+  ) -> Result<Self, crate::Error> {
+    if let Ok(buf) = MagicBuffer::from_v8(scope, value) {
+      return Ok(Self::Buffer(buf));
+    } else if let Ok(s) = crate::from_v8(scope, value) {
+      return Ok(Self::String(s));
+    }
+    Err(Error::ExpectedBuffer)
   }
 }