--- a/serde_v8/tests/de.rs
+++ b/serde_v8/tests/de.rs
@@ -178,24 +178,24 @@ fn de_map() {
 fn de_string_or_buffer() {
   dedo("'hello'", |scope, v| {
     let sob: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
-    assert_eq!(sob.as_slice(), &[0x68, 0x65, 0x6C, 0x6C, 0x6F]);
+    assert_eq!(sob.as_ref(), &[0x68, 0x65, 0x6C, 0x6C, 0x6F]);
   });
 
   dedo("new Uint8Array([97])", |scope, v| {
     let sob: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
-    assert_eq!(sob.as_slice(), &[97]);
+    assert_eq!(sob.as_ref(), &[97]);
   });
 
   dedo("new Uint8Array([128])", |scope, v| {
     let sob: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
-    assert_eq!(sob.as_slice(), &[128]);
+    assert_eq!(sob.as_ref(), &[128]);
   });
 
   dedo(
     "(Uint8Array.from([0x68, 0x65, 0x6C, 0x6C, 0x6F]))",
     |scope, v| {
       let sob: serde_v8::StringOrBuffer = serde_v8::from_v8(scope, v).unwrap();
-      assert_eq!(sob.as_slice(), &[0x68, 0x65, 0x6C, 0x6C, 0x6F]);
+      assert_eq!(sob.as_ref(), &[0x68, 0x65, 0x6C, 0x6C, 0x6F]);
     },
   );
 }