```
Before:
test de_sob_buf_16kb      ... bench:         597 ns/iter (+/- 16)
test de_sob_buf_1b        ... bench:          90 ns/iter (+/- 3)
test de_sob_buf_1kb       ... bench:         169 ns/iter (+/- 3)
test de_sob_buf_512kb     ... bench:      23,829 ns/iter (+/- 441)
test de_sob_str_1kb       ... bench:         555 ns/iter (+/- 6)
test de_sob_str_6b        ... bench:          76 ns/iter (+/- 0)

After:
test de_sob_buf_16kb      ... bench:          32 ns/iter (+/- 1)
test de_sob_buf_1b        ... bench:          32 ns/iter (+/- 2)
test de_sob_buf_1kb       ... bench:          32 ns/iter (+/- 1)
test de_sob_buf_512kb     ... bench:          32 ns/iter (+/- 2)
test de_sob_str_1kb       ... bench:         465 ns/iter (+/- 14)
test de_sob_str_6b        ... bench:          48 ns/iter (+/- 1)
```